Splash Added in GoogleSignInActivty.kt
