+++ client/android/application/app/src/main/java/com/example/myapplication/GoogleSignInActivty.kt
@@ -4,7 +4,10 @@ import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import android.util.Log
+import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.example.myapplication.AuthResponse.AuthTask
@@ -28,11 +31,18 @@ import retrofit2.Callback
 import retrofit2.Response
 
 class GoogleSignInActivty : AppCompatActivity() {
+
     lateinit var mGoogleSignInClient: GoogleSignInClient
     lateinit var mGoogleSignInAccount: GoogleSignInAccount
+    var key : String? = null
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_google_sign_in_activty)
+        val handler = Handler(Looper.getMainLooper())
+        handler.postDelayed({
+            checkLogin()
+        },2000)
         val gso: GoogleSignInOptions =
             GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
                 .requestIdToken("1076096355894-tlqr133lhbep9c9n8jkv24msqri7pvb2.apps.googleusercontent.com")
@@ -45,6 +55,22 @@ class GoogleSignInActivty : AppCompatActivity() {
         gSignInBtn.setOnClickListener {
             signIn()
         }
+    }
+
+    private fun checkLogin() {
+        val sp : SharedPreferences = getSharedPreferences("LOGIN", MODE_PRIVATE)
+        key = sp.getString("login_key",null)
+
+        if (key!=null){
+            val intent = Intent(this,TopicsActivity::class.java)
+            intent.putExtra("token",key)
+            Toast.makeText(this,"Welcome Back",Toast.LENGTH_SHORT).show()
+            startActivity(intent)
+            finish()}
+        else{
+            gSignInBtn.visibility = View.VISIBLE
+        }
+
     }
 
     private fun signIn() {
@@ -83,6 +109,7 @@ class GoogleSignInActivty : AppCompatActivity() {
                     val intent = Intent(this@GoogleSignInActivty, TopicsActivity::class.java)
                     intent.putExtra("token", response.body()?.token)
                     startActivity(intent)
+                    finish()
                 }
 
                 override fun onFailure(call: Call<AuthSuccess>, t: Throwable) {
