+++ client/android/application/app/src/main/res/layout/activity_google_sign_in_activty.xml
@@ -1,16 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:gravity="center"
-    tools:context=".GoogleSignInActivty"
-    android:orientation="vertical">
+    android:background="@color/Theme"
+    android:orientation="vertical"
+    tools:context=".GoogleSignInActivty">
 
+    <ImageView
+        android:layout_width="100dp"
+        android:layout_height="100dp"
+       android:layout_alignLeft="@id/refreshText"
+        android:layout_above="@id/refreshText"
+        android:src="@drawable/ic_logo_masai" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:layout_centerHorizontal="true"
+        android:fontFamily="monospace"
+        android:text="refresh"
+        android:id="@+id/refreshText"
+        android:textColor="@color/white"
+        android:textSize="50sp" />
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/gSignInBtn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="10dp"/>
-</LinearLayout>
+        android:layout_alignParentBottom="true"
+        android:layout_marginStart="10dp"
+        android:layout_marginEnd="10dp"
+        android:layout_marginBottom="30dp"
+        android:padding="10dp"
+        android:visibility="invisible" />
+</RelativeLayout>
