--- client/android/application/app/src/main/res/layout/activity_splash.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/Theme"
-    android:orientation="vertical"
-    android:gravity="center"
-    tools:context=".activities.SplashActivity">
-<ImageView
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:src="@drawable/ic_logo_masai"
-
-    android:layout_margin="30dp"
-    />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:fontFamily="monospace"
-        android:text="refresh"
-        android:textColor="@color/white"
-        android:textSize="30sp" />
-</LinearLayout>
