+++ client/android/application/app/src/main/res/values/themes.xml
@@ -12,5 +12,17 @@
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">@color/Theme</item>
         <!-- Customize your theme here. -->
+        <item name="android:displayOptions">useLogo</item>
+    </style>
+    <style name="NoActionBar" parent="Theme.AppCompat.DayNight.NoActionBar" >
+        <item name="colorPrimary">@color/Theme</item>
+        <item name="colorPrimaryVariant">@color/Theme</item>
+        <item name="colorOnPrimary">@color/white</item>
+        <!-- Secondary brand color. -->
+        <item name="colorSecondary">@color/teal_200</item>
+        <item name="colorSecondaryVariant">@color/teal_700</item>
+        <item name="colorOnSecondary">@color/black</item>
+        <!-- Status bar color. -->
+        <item name="android:statusBarColor" tools:targetApi="l">@color/Theme</item>
     </style>
 </resources>
